third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/css-tables/tentative/colspan-redistribution-expected.txt

This is a testharness.js-based test.
PASS table 1
PASS table 2
FAIL table 3 assert_equals: 
<table data-expected-width="492">
  <tbody><tr>
    <td data-expected-width="46"></td>
    <td></td>
    <td>x</td>
  </tr>
  <tr>
    <td colspan="2" style="width:20%"><div style="width:100px">100px</div></td>
</tr></tbody></table>
width expected 46 but got 92
PASS table 4
PASS table 5
PASS table 6
PASS table 7
PASS table 8
FAIL table 9 assert_equals: 
<table data-expected-width="344">
  <tbody><tr>
    <td style="width:20px" data-expected-width="146"><div style="width:100px">20/100px</div></td>
    <td style="width:40px"><div style="width:100px">40/100px</div></td>
    <td style="width:20px">x</td>
  </tr>
  <tr>
    <td data-expected-width="300" colspan="2" style=""><div style="width:300px">300px min</div></td>
  </tr>
</tbody></table>
width expected 146 but got 100
PASS table 10
PASS table 11
PASS table 12
PASS table 13
PASS table 14
FAIL table 15 assert_equals: 
<table style="width:1px" data-expected-width="166">
  <tbody><tr>
    <td style="width:70px" data-expected-width="42">0</td>
    <td style="width:70px" data-expected-width="42">0</td>
    <td data-expected-width="50"><div style="width:50px">50</div></td>
  </tr>
  <tr>
    <td colspan="3"><div style="width:150px"></div>
    </td>
  </tr>
</tbody></table>
width expected 42 but got 49
PASS table 16
PASS table 17
PASS table 18
FAIL table 19 assert_equals: 
<table style="width:1px" data-expected-width="252">
  <tbody><tr>
    <td style="width:40%" data-expected-width="80"><div style="width:20px">40%</div></td>
    <td style="width:50px" data-expected-width="120"><div style="width:50px">50</div></td>
    <td data-expected-width="20" style="width:100px"><div style="width:20px">20</div></td>
  </tr>
  <tr>
    <td colspan="2" data-expected-width="208"><div style="width:208px">208px min</div></td>
  </tr>
</tbody></table>
width expected 80 but got 57
PASS table 20
PASS table 21
FAIL table 22 assert_equals: 
<table data-expected-width="532">
  <tbody><tr>
    <td colspan="2"><div style="width:208px">200</div></td>
    <td><div style="width:50px">50</div></td>
  </tr>
  <tr>
    <td><div style="width:50px">50</div></td>
    <td colspan="2"><div style="width:408px">400</div></td>
  </tr>
  <tr>
    <td data-expected-width="100">Auto</td>
    <td data-expected-width="267"><div style="width:50px">Auto</div></td>
    <td data-expected-width="133">Auto</td>
  </tr>
</tbody></table>
width expected 532 but got 482
FAIL table 23 assert_equals: 
<table data-expected-width="482">
  <tbody><tr>
    <td colspan="2"><div style="width:408px">400</div></td>
    <td><div style="width:50px">50</div></td>
  </tr>
  <tr>
    <td><div style="width:50px">50</div></td>
    <td colspan="2"><div style="width:208px">200</div></td>
  </tr>
  <tr>
    <td data-expected-width="200">Auto</td>
    <td data-expected-width="200"><div style="width:50px">Auto</div></td>
    <td data-expected-width="50">Auto</td>
  </tr>
</tbody></table>
width expected 482 but got 532
FAIL table 24 assert_equals: 
<table data-expected-width="740">
  <tbody><tr>
    <td colspan="3"><div style="width:316px">300</div></td>
    <td><div style="width:50px">50</div></td>
  </tr>
  <tr>
    <td><div style="width:50px">50</div></td>
    <td colspan="3"><div style="width:616px">600</div></td>
  </tr>
  <tr>
    <td data-expected-width="100">Auto</td>
    <td data-expected-width="240"><div style="width:50px">Auto</div></td>
    <td data-expected-width="240"><div style="width:50px">Auto</div></td>
    <td data-expected-width="120">Auto</td>
  </tr>
</tbody></table>
width expected 740 but got 690
Harness: the test ran to completion.
